Renomme des fichiers et des symboles
pour mieux distinguer Parser, Visitor et Service
de traitement des expressions pour les indicateurs
et la personnalisations.
