+++ backend/src/personnalisations/personnalisations.module.ts
@@ -2,13 +2,13 @@ import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
 import { CollectivitesModule } from '../collectivites/collectivites.module';
 import { PersonnalisationsController } from './controllers/personnalisations.controller';
-import ExpressionParserService from './services/expression-parser.service';
+import PersonnalisationsExpressionService from './services/personnalisations-expression.service';
 import PersonnalisationsService from './services/personnalisations-service';
 
 @Module({
   imports: [CollectivitesModule, AuthModule],
-  providers: [ExpressionParserService, PersonnalisationsService],
-  exports: [ExpressionParserService, PersonnalisationsService],
+  providers: [PersonnalisationsExpressionService, PersonnalisationsService],
+  exports: [PersonnalisationsExpressionService, PersonnalisationsService],
   controllers: [PersonnalisationsController],
 })
 export class PersonnalisationsModule {}
