+++ backend/src/referentiels/compute-score/scores.service.spec.ts
@@ -9,7 +9,7 @@ import CollectivitesService from '../../collectivites/services/collectivites.ser
 import { PersonnalisationReponsesPayload } from '../../personnalisations/models/get-personnalisation-reponses.response';
 import { PersonnalisationConsequencesByActionId } from '../../personnalisations/models/personnalisation-consequence.dto';
 import { caePersonnalisationRegles } from '../../personnalisations/models/samples/cae-personnalisation-regles.sample';
-import ExpressionParserService from '../../personnalisations/services/expression-parser.service';
+import PersonnalisationsExpressionService from '../../personnalisations/services/personnalisations-expression.service';
 import PersonnalisationsService from '../../personnalisations/services/personnalisations-service';
 import ConfigurationService from '../../utils/config/configuration.service';
 import { DatabaseService } from '../../utils/database/database.service';
@@ -43,7 +43,7 @@ describe('ReferentielsScoringService', () => {
         ScoresService,
         SnapshotsService,
         PersonnalisationsService,
-        ExpressionParserService,
+        PersonnalisationsExpressionService,
         GetReferentielService,
       ],
     })
